requirements.txt
pymeshlab
opencv-python
matplotlib
imageio
scikit-image
numba
open3d
trimesh
plyfile
tensorboard
timm
h5py
pytorch_lightning
ffmpeg-python
mmcv
mediapy
pyrender
json5
# For Video Rendering
numpy
typing_extensions
pillow
iopath
markupsafe
jinja2
scikit-learn
sympy
networkx